✨ feat: support AWS partitions in CloudWatch log group ARN construction
- Extract partition from caller's ARN to handle aws-cn, aws-us-gov partitions
- Create dedicated arn.go with partition-aware ARN utilities
- Replace hardcoded "aws" partition with dynamic extraction
- Ensure compatibility across all AWS regions and partitions
