+++ internal/ecs/logs.go
@@ -206,8 +206,14 @@ func TailServiceLogs(ctx context.Context, clients *AWSClients, cluster, service 
 		return nil, nil, fmt.Errorf("failed to get caller identity: %w", err)
 	}
 
-	// Construct the LogGroup ARN
-	logGroupArn := fmt.Sprintf("arn:aws:logs:%s:%s:log-group:%s", clients.Region, *identity.Account, logGroup)
+	// Extract partition from caller's ARN to handle different AWS partitions
+	partition, err := extractPartitionFromARN(*identity.Arn)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to extract partition from caller ARN: %w", err)
+	}
+
+	// Construct the LogGroup ARN with correct partition
+	logGroupArn := buildARN(partition, "logs", clients.Region, *identity.Account, fmt.Sprintf("log-group:%s", logGroup))
 
 	// Use LogStreamNamePrefixes to capture all streams for this service's containers
 	logStreamPrefixPattern := fmt.Sprintf("%s/%s/", logStreamPrefix, containerName)
